Reimplement Lad
It is now a .NET 6 application.  Its inline format has changed to accept
a valid C# file.  That file must contain only one method with a switch
statement.  That method's parameters are removed in the output so it
will not conflict if the source file is part of the project.  The class
may also contain string fields that are interpreted as named
expressions.