Use provided implementations of scanner functions
The generated methods now invoke the provided ones.  This prevents
duplicated code.  This will also change the parameter and return value
requirements.  The provided methods must now include the pattern as the
first parameter and use it in the switch statement instead of the token
value since that will contain the token value.  The return value must be
nullable so the invoking method knows whether to loop or return.

This applies only to the inline implementation.  Lex has no such methods
so its implementation is not affected.